chore: add templates package
