+++ .storybook/modes.ts
@@ -1,21 +1,4 @@
+import { allModes } from "@hitachivantara/internal";
+
 /** These are are the themes returned by getThemesList */
-export const allModes = {
-  "Pentaho+ dawn": {
-    theme: "pentahoPlus dawn",
-  },
-  "Pentaho+ wicked": {
-    theme: "pentahoPlus wicked",
-  },
-  "DS5 dawn": {
-    theme: "ds5 dawn",
-  },
-  "DS5 wicked": {
-    theme: "ds5 wicked",
-  },
-  "DS3 dawn": {
-    theme: "ds3 dawn",
-  },
-  "DS3 wicked": {
-    theme: "ds3 wicked",
-  },
-};
+export { allModes };
