+++ apps/app/package.json
@@ -20,6 +20,7 @@
     "@hitachivantara/uikit-react-pentaho": "*",
     "@hitachivantara/uikit-react-viz": "*",
     "@hitachivantara/uikit-styles": "*",
+    "@hv/templates": "*",
     "@mui/material": "^5.14.20",
     "@phosphor-icons/react": "^2.1.7",
     "arquero": "^5.3.0",
@@ -28,7 +29,6 @@
     "json5": "^2.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-google-charts": "^4.0.1",
     "react-grid-layout": "^1.4.4",
     "react-router": "^7.1.5",
     "reactflow": "^11.10.1",
