+++ apps/docs/package.json
@@ -24,7 +24,8 @@
     "react-dom": "^18.3.1",
     "react-element-to-jsx-string": "^15.0.0",
     "react-live-runner": "^1.0.7",
-    "react-markdown": "^9.0.3"
+    "react-markdown": "^9.0.3",
+    "usehooks-ts": "^2.9.1"
   },
   "devDependencies": {
     "@hitachivantara/uikit-uno-preset": "*",
