+++ package.json
@@ -99,7 +99,6 @@
     "html-react-parser": "^3.0.16",
     "husky": "^8.0.3",
     "jsdom": "^26.1.0",
-    "leaflet": "^1.9.4",
     "lerna": "^8.1.7",
     "lint-staged": "^16.1.0",
     "npm-run-all2": "^8.0.4",
@@ -108,9 +107,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.11",
-    "react-google-charts": "^4.0.1",
     "react-hook-form": "^7.49.0",
-    "react-leaflet": "^4.2.1",
     "reactflow": "^11.10.1",
     "rimraf": "^6.0.1",
     "storybook": "^9.0.0",
@@ -118,7 +115,6 @@
     "typescript": "^5.8.2",
     "uid": "^2.0.2",
     "unocss": "^66.0.0",
-    "usehooks-ts": "^2.9.1",
     "vite": "^6.2.2",
     "vite-plugin-dts": "^4.5.4",
     "vitest": "^3.0.9",
@@ -151,6 +147,7 @@
   },
   "workspaces": [
     "packages/*",
-    "apps/*"
+    "apps/*",
+    "templates"
   ]
 }
