+++ packages/internal/src/utils.ts
@@ -1,5 +1,25 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-import { allModes } from ".../../../.storybook/modes";
+
+export const allModes = {
+  "Pentaho+ dawn": {
+    theme: "pentahoPlus dawn",
+  },
+  "Pentaho+ wicked": {
+    theme: "pentahoPlus wicked",
+  },
+  "DS5 dawn": {
+    theme: "ds5 dawn",
+  },
+  "DS5 wicked": {
+    theme: "ds5 wicked",
+  },
+  "DS3 dawn": {
+    theme: "ds3 dawn",
+  },
+  "DS3 wicked": {
+    theme: "ds3 wicked",
+  },
+};
 
 /**
  * Builds the Chromatic object needed to enable snapshots in Storybook stories
